fix bulkcopy
